requirements.txt
azure-storage
azure-mgmt-network
azure-mgmt-resource
boto3
botocore
cerberus
docopt
google-api-python-client
requests
retrying
tox
pyinstaller==3.2
py
pytest
pyyaml
